Bug fix of first implementation of ModifyWanteds command
